remove unused testing helper func & mv integration conf fixt to parent test dir
